requirements.txt
streamlit
pydub
mistralai
python-dotenv
requests
groq
